Move test cases from CustomObjectServiceImplIT to CustomObjectServiceImplTest
